Crate coro

This crate implements coroutines in rust. Please note that, in this implementation, one cannot yield a function inside some control flow structre. Instead, the function must return either a Yield or a Done. Note also that a coroutine may not be an actual function. The coroutine only has to implement the trait Coroutine<T>. Actually, functions by themselves are not coroutines. Instead, they must be wrapped with CoFn or a custom wrapper. Look at this example.

extern crate coro;

use coro::{CoFn, CoTasking, Yield, Done};

#[derive(Debug, Clone, PartialEq, Eq)]
enum Msg {
    Empty,
    Push(i64),
    Return(Vec<i64>),
}

let consumer = CoFn::new(|scheduler| {
    let producer = scheduler.id_for("producer").unwrap();
    let mut rtrn = Vec::new();
    move |message| match message {
        Msg::Push(x) => {
            rtrn.push(x);
            Yield(producer, Msg::Empty)
        },
        _ => Done(Msg::Return(rtrn.clone()))
    }
});

let producer = CoFn::new(|scheduler| {
    let consumer = scheduler.id_for("consumer").unwrap();
    let mut source = vec![232, -1232, 43];
    move |_| match source.pop() {
        Some(x) => Yield(consumer, Msg::Push(x)),
        _ => Yield(consumer, Msg::Empty),
    }
});

let result = CoTasking::new()
    .task("producer", producer)
    .task("consumer", consumer)
    .run("producer", Msg::Empty);
assert_eq!(result, Msg::Return(vec![43, -1232, 232]));

Re-exports

pub use self::Pause::*;

Structs

CoFn
A boilerplate reduction for corroutines implemented as regular rust functions.

CoId
Identifies a single coroutine instance in a scheduler.

CoTasking
Builds a Co-operative tasking with some given coroutines.

Scheduler
A type that manages a co-operative tasking execution. Mostly invisible.

Enums

Pause
Must be returned from a single call of resume.

Traits

Coroutine
This trait defines a coroutine in order to be used with this crate.